refactor(ner-plugin): classes renaming and reorganization
